spike(#351): move solver to src/, wire into compute_layout (env-gated) [skip ci]
Architecture move from PR #350's pure post-pass: solver now lives in
src/nf_metro/layout/constraint_solver_v2.py and runs unconditionally at
the end of compute_layout (gated on NF_METRO_NO_SOLVER=1 to let diff
scripts run the engine without the override).

The render-diff CI on this branch will fire the solver on every gallery
fixture, against main's pure-engine baseline. That's the authoritative
visual review per the issue.

scratch/constraint_spike_v3_model.py is now a thin wrapper that calls
compute_layout. scratch/constraint_spike_v3_compare.py runs the env-var
toggle to compare engine vs solver locally.

ruff check + format clean.

Refs #351.
